+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
-# [1.2.4-3-rc.1]
+# [1.2.4-3-final]
 
 - Fixed map crash after airports procedures have been opened, but nothing was selected
 - Fixed "show pilot stats" toggle not being saved
+- Add day/night layer (terminator) to "Map layers" settings. More customization options will be available later
+- Add LECB Area mapping to ATC duplication settings by 1558357
+- Add ZNY and N90 to ATC Duplication by 1590802
 
 # [1.2.4-2-final]